type_convertion_bug

+++ sft_dpo/sft.py
@@ -33,8 +33,8 @@ class SFT:
 
         trainer = SFTTrainer(
             model=model, 
-            train_dataset=dataset[:len(dataset) * self.proportion], 
-            eval_dataset=dataset[len(dataset) * self.proportion: len(dataset) * (self.proportion + 0.2)],
+            train_dataset=dataset[:int(len(dataset) * self.proportion)], 
+            eval_dataset=dataset[int(len(dataset) * self.proportion): int(len(dataset) * (self.proportion + 0.2))],
             peft_config=lora_config,
             tokenizer=tokenizer,
             max_seq_length=self.config["max_seq_length"],
